script for testing phnological ndvi chain added (for Vegetation Status processor). BandsExtractor uses the provided resolution from the input image (update). ResampleAtS2Res is using now only the provided output

+++ sen2agri-processors/CropType/BandsExtractor/BandsExtractor.cpp
@@ -1267,7 +1267,7 @@ private:
            m_ResamplersList->PushBack(resampler);
            return resampler;
        }
-       const float ratio = curRes / wantedRes;
+       const float ratio = (float)curRes / (float)wantedRes;
 
        // Set the interpolator
        LinearInterpolationType::Pointer interpolator = LinearInterpolationType::New();
